+++ tl-3d-threejs/src/main/java/com/top_logic/threed/threejs/scene/PartNode.java
@@ -69,13 +69,13 @@ public class PartNode extends SceneNode {
 	}
 
 	@Override
-	public com.top_logic.threed.threejs.scene.PartNode setTransform(java.util.List<? extends Float> value) {
+	public com.top_logic.threed.threejs.scene.PartNode setTransform(java.util.List<? extends Double> value) {
 		internalSetTransform(value);
 		return this;
 	}
 
 	@Override
-	public com.top_logic.threed.threejs.scene.PartNode addTransform(float value) {
+	public com.top_logic.threed.threejs.scene.PartNode addTransform(double value) {
 		internalAddTransform(value);
 		return this;
 	}
